feat: Move Request match button / remove unnecessary component
