+++ src/pages/student/MatchingStudent.tsx
@@ -125,7 +125,7 @@ const MatchingStudent: React.FC<Props> = () => {
         // eslint-disable-next-line react-hooks/exhaustive-deps
     }, []);
 
-    const matchRequestCount = data?.me?.student?.openMatchRequestCount;
+    const matchRequestCount = data?.me?.student?.openMatchRequestCount ?? 0;
 
     return (
         <AsNavigationItem path="matching">
@@ -144,17 +144,6 @@ const MatchingStudent: React.FC<Props> = () => {
                         <Heading paddingBottom={space['0.5']}>{t('matching.request.check.title')}</Heading>
                         <VStack space={space['0.5']}>
                             <Text paddingBottom={space['0.5']}>{t('matching.request.check.content')}</Text>
-                            {(data?.me?.student?.canRequestMatch.allowed && (
-                                <Button width={ButtonContainer} marginBottom={space['1.5']} onPress={() => navigate('/request-match')}>
-                                    {t('dashboard.helpers.buttons.requestMatchStudent')}
-                                </Button>
-                            )) || (
-                                <AlertMessage
-                                    content={t(
-                                        `lernfair.reason.matching.tutor.${data?.me?.student?.canRequestMatch?.reason}` as unknown as TemplateStringsArray
-                                    )}
-                                />
-                            )}
                         </VStack>
                         <NavigationTabs
                             tabs={[
@@ -166,7 +155,7 @@ const MatchingStudent: React.FC<Props> = () => {
                                     title: (
                                         <span style={{ display: 'flex' }}>
                                             {t('matching.request.check.tabs.tab2')}{' '}
-                                            {matchRequestCount && matchRequestCount > 0 && (
+                                            {matchRequestCount > 0 && (
                                                 <Circle bgColor="danger.500" size="5">
                                                     <Text fontSize="xs" color="white">
                                                         {matchRequestCount}
@@ -218,6 +207,19 @@ const MatchingStudent: React.FC<Props> = () => {
                                 },
                             ]}
                         />
+                        <VStack space={['0.5']}>
+                            {(data?.me?.student?.canRequestMatch.allowed && (
+                                <Button width={ButtonContainer} marginBottom={space['1.5']} onPress={() => navigate('/request-match')}>
+                                    {t('dashboard.helpers.buttons.requestMatchStudent')}
+                                </Button>
+                            )) || (
+                                <AlertMessage
+                                    content={t(
+                                        `lernfair.reason.matching.tutor.${data?.me?.student?.canRequestMatch?.reason}` as unknown as TemplateStringsArray
+                                    )}
+                                />
+                            )}
+                        </VStack>
                         {process.env.REACT_APP_HOMEWORKHELP !== '' && (
                             <VStack space={space['0.5']} paddingX={space['1']} width="100%" marginX="auto" maxWidth={ContainerWidth}>
                                 <Heading paddingBottom={space['0.5']}>{t('matching.homeworkhelp.title')}</Heading>
